requirements.txt
# Frida Orchestrator Backend
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
google-generativeai==0.8.0
rembg==2.0.59
pillow==10.4.0
jinja2==3.1.4
weasyprint==62.3
python-dotenv==1.0.1
supabase==2.7.0
